stock-analyzer-pro/backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
sqlalchemy==2.0.15
pandas==2.0.1
numpy==1.24.2
scikit-learn==1.2.2
aiohttp==3.8.4
websockets==11.0.3
requests==2.28.2
beautifulsoup4==4.12.2
transformers==4.28.1
torch==2.0.0
torchvision==0.15.1
talib-binary==0.4.25
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==4.5.4
psycopg2-binary==2.9.6
kafka-python==2.0.2
python-dotenv==1.0.0
pydantic==1.10.7
arch==5.3.1
statsmodels==0.13.5
openai==0.27.4
scipy==1.10.1
